+++ python-wrapper/src/neo4j_viz/__init__.py
@@ -5,8 +5,11 @@ from .options import (
     ForceDirectedLayoutOptions,
     HierarchicalLayoutOptions,
     Layout,
+    NvlOptions,
     Packing,
+    PanPosition,
     Renderer,
+    WidgetOptions,
 )
 from .relationship import Relationship
 from .visualization_graph import VisualizationGraph
@@ -15,6 +18,9 @@ from .widget import GraphWidget
 __all__ = [
     "VisualizationGraph",
     "GraphWidget",
+    "WidgetOptions",
+    "NvlOptions",
+    "PanPosition",
     "Node",
     "Relationship",
     "CaptionAlignment",